+++ elements/src/components/pos-switch/pos-case/pos-case.tsx
@@ -1,0 +1,26 @@
+import { Component, Element, State } from '@stencil/core';
+
+/**
+ * Defines a template to use if the specified condition is met - to be used with [pos-switch](https://pod-os.org/reference/elements/components/pos-switch/).
+ * See [storybook](https://pod-os.github.io/PodOS/storybook/?path=/story/basics--pos-switch) for an example.
+ */
+@Component({
+  tag: 'pos-case',
+  shadow: false,
+})
+export class PosCase {
+  @Element() host: HTMLElement;
+  @State() error: string = null;
+
+  componentWillLoad() {
+    const templateElement = this.host.querySelector('template');
+    if (templateElement == null) {
+      this.error = 'No template element found';
+    }
+  }
+
+  render() {
+    if (this.error) return this.error;
+    return null;
+  }
+}